% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-globals.R
\docType{package}
\name{gwas2crispr-package}
\alias{gwas2crispr}
\alias{gwas2crispr-package}
\title{gwas2crispr package-level imports}
\description{
Provides a reproducible pipeline to conduct genome-wide association studies (GWAS) and extract single-nucleotide polymorphisms (SNPs) for a human trait or disease. Given aggregated GWAS dataset(s) and a user-defined significance threshold, the package retrieves significant SNPs from the GWAS Catalog and the Experimental Factor Ontology (EFO), annotates their gene context, and can write a harmonised metadata table in comma-separated values (CSV) format, genomic intervals in the Browser Extensible Data (BED) format, and sequences in the FASTA (text-based sequence) format with user-defined flanking regions for clustered regularly interspaced short palindromic repeats (CRISPR) guide design. For details on the resources and methods see: Buniello et al. (2019) \doi{10.1093/nar/gky1120}; Sollis et al. (2023) \doi{10.1093/nar/gkac1010}; Jinek et al. (2012) \doi{10.1126/science.1225829}; Malone et al. (2010) \doi{10.1093/bioinformatics/btq099}; Experimental Factor Ontology (EFO) \url{https://www.ebi.ac.uk/efo}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leopard0ly/gwas2crispr}
  \item Report bugs at \url{https://github.com/leopard0ly/gwas2crispr/issues}
}

}
\author{
\strong{Maintainer}: Othman S. I. Mohammed \email{admin@leopard.ly}

Other contributors:
\itemize{
  \item LEOPARD.LY LTD [copyright holder]
}

}
\keyword{internal}
